requirements.txt
aiogram
aiohttp
aiosqlite
apscheduler
python-dotenv
